staging: updated data types

+++ models/staging/stg_cities.sql
@@ -4,13 +4,13 @@ with source as(
 ),
 renamed as(
     select
-        cast(location_id as integer) as location_id,
-        cast(latitude as float64) as latitude,
-        cast(longitude as float64) as longitude,
-        cast(elevation as float64) as elevation,
+        cast(location_id    as integer) as location_id,
+        cast(latitude       as float64) as latitude,
+        cast(longitude      as float64) as longitude,
+        cast(elevation      as float64) as elevation,
         case 
-            when (cast(latitude as float64) = 55.75 and cast(longitude as float) = -4.25) then 'Glasgow'
-            when (cast(latitude as float64) = 25 and cast(longitude as float) = 55.25) then 'Dubai'
+            when (cast(latitude as float64) = 55.75 and cast(longitude as float64) = -4.25) then 'Glasgow'
+            when (cast(latitude as float64) = 25    and cast(longitude as float64) = 55.25) then 'Dubai'
             else 'Cairo'
             end as city_name
     from source
